+++ src/obstacle.c
@@ -6,13 +6,13 @@ obstacle_sensors_t obstacle_sensors;
 
 // Constants for the calculation of the range finders
 // a = 1024 / 3.3V * 28
-const float a = 8688.4;
-const float b = 0.42;
+const double a = 8688.4;
+const double b = 0.42;
 
-inline int
+inline double
 convert_to_mm(int reading)
 {
-  return (int)(a / reading - b);
+  return (a / reading - b);
 }
 
 void
